Refactor CLI commands into submodule
